feat(core): at menu ux (#9045)
fix AF-1843
