+++ packages/frontend/i18n/src/resources/en.json
@@ -1492,9 +1492,11 @@
   "com.affine.attachment.preview.error.subtitle": "file type not supported.",
   "com.affine.pdf.page.render.error": "Failed to render page.",
   "com.affine.editor.journal-conflict.title": "Duplicate Entries in Today's Journal",
-  "com.affine.editor.at-menu.link-to-doc": "Link to Doc",
-  "com.affine.editor.at-menu.new-doc": "New Doc",
-  "com.affine.editor.at-menu.create-doc": "Create \"{{name}}\" Doc",
+  "com.affine.editor.at-menu.link-to-doc": "Search for \"{{query}}\"",
+  "com.affine.editor.at-menu.recent-docs": "Recent",
+  "com.affine.editor.at-menu.new-doc": "New",
+  "com.affine.editor.at-menu.create-page": "New \"{{name}}\" page",
+  "com.affine.editor.at-menu.create-edgeless": "New \"{{name}}\" edgeless",
   "com.affine.editor.at-menu.import": "Import",
   "com.affine.editor.at-menu.more-docs-hint": "{{count}} more docs",
   "com.affine.editor.at-menu.journal": "Journal",
